gutcheck-web/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
streamlit>=1.30.0
mistralai>=1.0.0
httpx>=0.25.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
pydantic>=2.5.0
